+++ config/index.js
@@ -1,5 +1,13 @@
 const config = {
-    port: 3000
+    development: {
+        port: 3000
+    },
+    test: {
+        port: 3001
+    },
+    production: {
+        port: 3003
+    }
 }
 
 module.exports = config;
